pyproject.toml
[project]
name = "git-archeologist"
version = "0.1.0"
description = "MCP Server for Semantic Git Archaeology using Gemini"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",         # הפרוטוקול המרכזי
    "fastmcp>=0.1.0",          # ספרייה שמקצרת תהליכים בבניית השרת
    "httpx>=0.27.0",           # לביצוע קריאות HTTP (אם נצטרך)
    "GitPython>=3.1.43",       # הספרייה שחופרת בתוך ה-Git שלך
    "google-generativeai>=0.8.0", # החיבור ל-Gemini API
    "python-dotenv>=1.0.1",    # לניהול ה-API Key בצורה מאובטחת
    "google-genai>=1.56.0",
    "pytest>=9.0.2",
]

[tool.uv]
managed = true
